requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-debug-toolbar==4.2.0
django-silk==5.0.4
djangorestframework==3.14.0
google==3.0.0
googlemaps==4.10.0
gprof2dot==2022.7.29
idna==3.4
mysqlclient==2.2.0
Pillow==10.0.1
pycodestyle==2.11.1
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
